refactor: consolidate types organization
- Move shared business types to lib/types.ts\n- Create dedicated params.ts for API parameters\n- Remove duplicate type definitions\n- Keep user and starknet types separate\n- Update imports across affected files
